# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dungeon/attempts:
    post:
      operationId: createDungeonAttempt
      tags: [dungeon]
      summary: Register an authenticated browser-played dungeon attempt
      description: Attempts are idempotent per account and requestId, have a sixty-attempt hourly abuse safeguard, and cannot start after the account completes that cycle's encounter.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DungeonAttemptInput" }
      responses:
        "201":
          description: New attempt registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DungeonAttempt" }
        "200":
          description: Idempotent replay of an existing attempt
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DungeonAttempt" }
        "401": { description: Authentication required }
        "409": { description: The current hourly encounter is already complete }
        "429": { description: Attempt limit reached for the current hourly cycle }
  /dungeon/attempts/{attemptId}/finish:
    post:
      operationId: finishDungeonAttempt
      tags: [dungeon]
      summary: Record the browser-reported result and award a single hourly completion buff on a qualifying win
      parameters:
        - in: path
          name: attemptId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DungeonAttemptFinishInput" }
      responses:
        "200":
          description: Idempotent settlement receipt
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DungeonAttemptFinishReceipt" }
        "401": { description: Authentication required }
        "404": { description: Attempt not found for the signed-in account }
        "409": { description: A win cannot be settled before the minimum play duration }
...
  /arena/profiles:
    get:
      operationId: searchArenaProfiles
      tags: [arena]
      summary: Search signed-in players' public Arena profiles
      parameters:
        - $ref: "#/components/parameters/RankedTeamSize"
        - in: query
          name: search
          required: false
          schema: { type: string, maxLength: 64 }
      responses:
        "200":
          description: Matching player profiles
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ArenaProfileCard" }
  /arena/profiles/{playerId}:
    get:
      operationId: getArenaPublicProfile
      tags: [arena]
      summary: View a player's Arena profile and featured replays
      parameters:
        - $ref: "#/components/parameters/RankedTeamSize"
        - in: path
          name: playerId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Public profile, with private recent matches when viewing your own profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArenaPublicProfile" }
...
  /arena/notifications/read:
    post:
      operationId: markArenaNotificationsRead
      tags: [arena]
      summary: Dismiss only the notifications the player has seen
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ArenaNotificationsReadInput" }
      responses:
        "200":
          description: Updated count
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArenaNotificationsReadReceipt" }
...
  /arena/challenges/{challengeId}/feature:
    post:
      operationId: featureArenaMatch
      tags: [arena]
      summary: Save a completed recording on your profile (up to five)
      parameters:
        - in: path
          name: challengeId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArenaFeatureReceipt" }
    delete:
      operationId: unfeatureArenaMatch
      tags: [arena]
      summary: Remove a saved recording from your profile
      parameters:
        - in: path
          name: challengeId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArenaFeatureReceipt" }
...
components:
  parameters:
    RankedTeamSize:
      in: query
      name: teamSize
      required: false
      schema: { type: integer, enum: [2, 3] }
  schemas:
...
    ArenaProfileCard:
      type: object
      required: [playerId, name, rating, wins, losses, defensePublished]
      properties:
        playerId: { type: string }
        name: { type: string }
        rating: { type: integer }
        wins: { type: integer }
        losses: { type: integer }
        defensePublished: { type: boolean }
    ArenaFeaturedMatch:
      type: object
      required: [id, opponentId, opponentName, outcome, ratingDelta, createdAt, recordingContentType, featured]
      properties:
        id: { type: string, format: uuid }
        opponentId: { type: string }
        opponentName: { type: string }
        outcome: { type: string, enum: [win, loss, draw] }
        ratingDelta: { type: integer }
        createdAt: { type: string, format: date-time }
        recordingContentType: { type: string }
        featured: { type: boolean }
    ArenaPublicProfile:
      allOf:
        - $ref: "#/components/schemas/ArenaProfileCard"
        - type: object
          required: [isOwnProfile, defenseTeam, featuredMatches, recentMatches]
          properties:
            isOwnProfile: { type: boolean }
            ranked2v2: { $ref: "#/components/schemas/ArenaRankedScore" }
            ranked3v3: { $ref: "#/components/schemas/ArenaRankedScore" }
            defenseTeam:
              type: [object, "null"]
              description: Public portrait lineup from the currently published defense; private builds are never included.
              required: [heroes, captainClass]
              properties:
                heroes:
                  type: array
                  minItems: 2
                  maxItems: 3
                  items:
                    type: object
                    required: [classId]
                    properties:
                      classId: { type: string, enum: [frostmage, priest, warrior, rogue, paladin, archer, warlock, druid, shaman] }
                captainClass: { type: [string, "null"] }
            featuredMatches:
              type: array
              items: { $ref: "#/components/schemas/ArenaFeaturedMatch" }
            recentMatches:
              type: array
              description: Only included for the owner; never expose unfeatured replays to other players.
              items: { $ref: "#/components/schemas/ArenaFeaturedMatch" }
    ArenaRankedScore:
      type: object
      required: [rating, wins, losses]
      properties:
        rating: { type: integer }
        wins: { type: integer }
        losses: { type: integer }
    ArenaFeatureReceipt:
      type: object
      required: [featured]
      properties:
        featured: { type: boolean }
...
    ArenaNotificationsReadInput:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          minItems: 1
          maxItems: 100
          uniqueItems: true
          items: { type: string, format: uuid }
    ArenaNotificationsReadReceipt:
      type: object
      required: [markedCount]
      properties:
        markedCount: { type: integer }
...
    DungeonEncounterId:
      type: string
      enum: [frost, demon, temple]
    DungeonBuff:
      type: object
      required: [id, cycle, kind, value, expiresAt, name, icon]
      properties:
        id: { $ref: "#/components/schemas/DungeonEncounterId" }
        cycle: { type: integer }
        kind: { type: string, enum: [maxHp, damage, healing] }
        value: { type: number }
        expiresAt: { type: integer, description: Epoch milliseconds at the next hourly reset. }
        name: { type: string }
        icon: { type: string }
    DungeonStatus:
      type: object
      required: [serverNow, cycle, resetAt, encounterId, completed, activeBuff]
      properties:
        serverNow: { type: integer, description: Current server epoch milliseconds. }
        cycle: { type: integer, description: Floor of serverNow divided by one hour in milliseconds. }
        resetAt: { type: integer, description: Epoch milliseconds at the next hourly reset. }
        encounterId: { $ref: "#/components/schemas/DungeonEncounterId" }
        completed: { type: boolean }
        activeBuff:
          oneOf:
            - $ref: "#/components/schemas/DungeonBuff"
            - type: "null"
    DungeonAttemptInput:
      type: object
      required: [requestId, heroes]
      properties:
        requestId: { type: string, format: uuid }
        heroes:
          type: array
          minItems: 2
          maxItems: 3
          items: { type: string }
    DungeonAttempt:
      type: object
      required: [id, cycle, encounterId, startedAt, status]
      properties:
        id: { type: string, format: uuid }
        cycle: { type: integer }
        encounterId: { $ref: "#/components/schemas/DungeonEncounterId" }
        startedAt: { type: integer, description: Epoch milliseconds. }
        status: { $ref: "#/components/schemas/DungeonStatus" }
    DungeonAttemptFinishInput:
      type: object
      required: [outcome]
      properties:
        outcome: { type: string, enum: [win, loss, abandoned] }
    DungeonAttemptFinishReceipt:
      type: object
      required: [status, awarded]
      properties:
        status: { $ref: "#/components/schemas/DungeonStatus" }
        awarded: { type: boolean }
        gold: { type: integer }
        balance: { type: integer }
...